fix issue when a perm banned player gets banned again

+++ src/main/java/de/strifel/vbans/commands/CommandBan.java
@@ -59,7 +59,7 @@ public class CommandBan implements Command {
                     if (uuid != null) {
                         if (!Util.hasOfflineProtectBanPermission(uuid, vBans) || commandSource instanceof ConsoleCommandSource) {
                             Ban currentBan = database.getBan(uuid);
-                            if (currentBan != null && commandSource.hasPermission("VBans.ban.topPerm")) {
+                            if (currentBan != null && commandSource.hasPermission("VBans.ban.topPerm") && currentBan.getUntil() != -1) {
                                 database.purgeActiveBans(uuid, commandSource instanceof ConsoleCommandSource ? "Console" : ((Player) commandSource).getUniqueId().toString());
                                 commandSource.sendMessage(TextComponent.of(strings[0] + " was already banned until " + Util.UNBAN_DATE_FORMAT.format(currentBan.getUntil() * 1000) + ". I removed that and created a perm ban.").color(TextColor.YELLOW));
                                 currentBan = null;
@@ -68,7 +68,7 @@ public class CommandBan implements Command {
                                 database.addBan(uuid, -1, commandSource instanceof ConsoleCommandSource ? "Console" : ((Player) commandSource).getUniqueId().toString(), reason);
                                 commandSource.sendMessage(TextComponent.of("You banned " + strings[0]).color(TextColor.YELLOW));
                             } else {
-                                commandSource.sendMessage(TextComponent.of(strings[0] + " is already banned until " + Util.UNBAN_DATE_FORMAT.format(currentBan.getUntil() * 1000)).color(TextColor.RED));
+                                commandSource.sendMessage(TextComponent.of(strings[0] + " is already banned until " + (currentBan.getUntil() == -1 ? "the end of his life." : Util.UNBAN_DATE_FORMAT.format(currentBan.getUntil() * 1000))).color(TextColor.RED));
                             }
                         } else {
                             commandSource.sendMessage(TextComponent.of("You are not allowed to ban this player!").color(TextColor.RED));
